18.7 So with our update policy finished, we can just use this as a basis for everything else. And for example, for deleting, all we need to do is check if the token has the ability to delete the ticket or delete own ticket. All of the other logic is going to be the same.  And the replace is going to be a lot simpler because the typical user does not have the ability to replace the ticket. So this will be a little bit different in that we will check if the token has the replace ticket ability, and if so, then we return true. Otherwise, we return false.  And really creating is going to be essentially the same thing. Let's call this store just because we're keeping it in line with the convention of being the name of the method on the controller. So we have delete, replace, store goes next. The ability is create ticket, and that should be fine.  So we should be able to go to the ticket controller and just apply these policies. So let's see, for the store method, we are finding that user. I guess I took out the code to actually create that ticket. We need to add that back in.  But for right now, what I'm concerned with is to use the Gate::authorize. Okay, so yeah, so we will check to see if the user can store this. We will pass in null here because we don't have a ticket. But then let's add the to do, create ticket there. Let's scroll on down to the replace. So we fetch the ticket before we update that. We need to check our policy for replace, and then essentially the same thing for delete. So once again, check our policy, but we are checking the delete policy, and we should be good to go.  And one of the really nice things about using token abilities is that we can be as granular as we need, for example, for updating a ticket. Managers should be able to update everything about a ticket, but a typical user should only be able to update certain fields. Because otherwise, they would be able to replace everything, and we don't want to provide that functionality to them.  So in the next episode, we will look at how we can do that.

+++ app/Http/Controllers/Api/v1/TicketController.php
@@ -28,6 +28,10 @@ class TicketController extends ApiController {
 	public function store(StoreTicketRequest $request) {
 		try {
 			$user = User::findOrFail($request->input('data.relationships.author.data.id'));
+
+			Gate::authorize('store', null);
+
+			//To do create ticket
 		} catch (ModelNotFoundException $exception) {
 			return $this->ok('User Not Found', [
 				'error' => 'The provided user id does not exists'
@@ -58,12 +62,14 @@ class TicketController extends ApiController {
 		try {
 			$ticket = Ticket::findOrFail($ticket_id);
 
+			Gate::authorize('replace', $ticket);
+
 			$ticket->update($request->mappedAttributes());
 
 			return new TicketResource($ticket);
 		} catch (ModelNotFoundException $exception) {
 			return  $this->error('Ticket Not Found', 404);
-		}
+		} 
 	}
 
 	/**
@@ -92,6 +98,9 @@ class TicketController extends ApiController {
 	public function destroy($ticket_id) {
 		try {
 			$ticket = Ticket::findOrFail($ticket_id);
+
+			Gate::authorize('delete', $ticket);
+
 			$ticket->delete();
 
 			return $this->ok('Ticket successfully deleted');
